Blower投げ数制限実装

+++ Source/Cutter/Private/TableRow/StageRowData.h
@@ -26,7 +26,7 @@ public:
 	float limitTime = {};
 	
 	UPROPERTY(EditAnywhere, BlueprintReadOnly)
-	float blowerCount = {};
+	int32 blowerCount = {};
 	
 	UPROPERTY(EditAnywhere, BlueprintReadOnly)
 	TObjectPtr<UDataTable> obstacleSpawnData = {};
